Made listeners lazy-loaded
This change makes the listeners lazy-loaded to enhance performances.

Caution: for ORM listeners the id in the config file is now the id of
the DBAL connection instead of the entity manager (this means all the
entity managers using the same connection will use the same behaviors).

the locale of the session is not set automatically with this
implementation (at the moment).