Refactor the registration of services
Services are now split in separate files per extension, which are loaded
only when the extension is enabled. This simplifies the DI extension as it
does not need to add event_subscriber tags conditionally on services when the
corresponding extensions are used.
The processing logic is now also shared between the ORM and the ODM
registration, ensuring that any listener priority being applied is done for
both of them (the ODM was missing some of them).